refactor(cli): extract quality embedding helpers to _quality_embedding
B2 of the Round 2 plan. Moves the four SentenceTransformer-based
helpers (_load_sentence_embedder, _cosine,
embedding_agreement_for_asset, semantic_grounding_score) out of
amx/cli_support/quality.py into a focused
amx/cli_support/_quality_embedding.py.

Tier 1 of the quality framework lives here - pairwise cosine
agreement + semantic grounding via all-MiniLM-L6-v2. Pure functions
with the sentence-transformers dependency lazy-imported on first
use.

quality.py re-exports each name. quality.py: 1020 -> 944 LOC (-76).
